pyproject.toml
[tool.poetry]
name = "cmem-plugin-parameters"
version = "0.0.0"
license = "Apache-2.0"
description = "eccenca Corporate Memory Workflow plugin for parameter input"
authors = ["Thomas Wilmering <thomas.wilmering@eccenca.com>"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Plugins",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
readme = "README-public.md"
keywords = [
    "eccenca Corporate Memory", "plugin"
]

[tool.poetry.dependencies]
python = "^3.9"
cmem-plugin-base = "^2.1.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
black = "^22.1.0"
bandit = "^1.7.2"
pylint-junit = "^0.3.2"
safety = "^1.10.3"
genbadge = "^1.0.6"
flake8-formatter-junit-xml = "^0.0.6"
typed-ast = "^1.5.2"
mypy = "^0.931"
coverage = "^6.3.2"
defusedxml = "^0.7.1"

[tool.poetry.group.dev.dependencies]
types-PyYAML = "^6.0.12"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
dirty = true

[tool.pylint.MASTER]
load-plugins="pylint_junit"

[tool.pylint.General]
ignore = "version.py"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
disable = "fixme"

# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#pylint
[tool.pylint.messages_control]
disable = "R0903"

[tool.pylint.format]
max-line-length = "88"

[tool.mypy]
warn_return_any = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = ""

